feat(ui): refine dashboard and unify level badges across admin
