+++ apps/web/src/index.css
@@ -134,6 +134,7 @@ select {
   position: relative;
   isolation: isolate;
   overflow: hidden;
+  --ui-hover-border-color: rgba(212, 162, 76, 0.9);
   transition:
     transform 260ms cubic-bezier(0.22, 1, 0.36, 1),
     box-shadow 260ms cubic-bezier(0.22, 1, 0.36, 1),
@@ -141,6 +142,19 @@ select {
     background-color 260ms ease,
     filter 260ms ease;
   will-change: transform, box-shadow;
+}
+
+.ui-surface-hover::before {
+  content: "";
+  position: absolute;
+  inset: 0;
+  border-radius: inherit;
+  pointer-events: none;
+  box-shadow: inset 0 0 0 0 rgba(212, 162, 76, 0);
+  transition:
+    box-shadow 260ms cubic-bezier(0.22, 1, 0.36, 1),
+    opacity 260ms ease;
+  opacity: 0.92;
 }
 
 .ui-surface-hover::after {
@@ -173,11 +187,19 @@ select {
   will-change: transform;
 }
 
+.ui-surface-hover--no-accent {
+  --ui-hover-border-color: transparent;
+}
+
 @media (hover: hover) and (pointer: fine) {
   .ui-surface-hover:hover {
     transform: translateY(-6px) scale(1.008);
     box-shadow: 0 26px 46px -32px rgba(15, 23, 42, 0.24);
     filter: saturate(1.02);
+  }
+
+  .ui-surface-hover:hover::before {
+    box-shadow: inset 0 0 0 1.5px var(--ui-hover-border-color);
   }
 
   .ui-surface-hover:hover::after {
@@ -268,6 +290,7 @@ select {
 @media (prefers-reduced-motion: reduce) {
   .ui-animate-in,
   .ui-surface-hover,
+  .ui-surface-hover::before,
   .ui-surface-hover::after,
   .ui-surface-hover__icon,
   .ui-surface-hover__bar,
